Feature/lh 1275 and lh 1350 fixes (#877)
* Add confirmation for bulk delete Assessor Assessment

* Make various report modals close

* Normalize translation keys

+++ app/frontend/modules/admin/modules/campaigns/routes/Campaign/routes/Participants/Assessors/AssessorDetails/AssessmentList/index.tsx
@@ -69,10 +69,21 @@ const AssessmentList: React.FC<Props> = ({
     fetch(parsedCampaignId, parsedAssessorId, tableConfig)
   }, [tableConfig])
 
-  const handleBulkDelete = async () => {
-    await bulkDelete(parsedCampaignId, parsedAssessorId, selectedIds)
-    await fetch(parsedCampaignId, parsedAssessorId, tableConfig)
-    message.success(I18n.t('administration.assessor.assessments.bulk_delete_successful'))
+  const handleBulkDelete = () => {
+    Modal.confirm({
+      title: I18n.t('common.text.confirm'),
+      icon: <ExclamationCircleOutlined />,
+      centered: true,
+      width: 650,
+      content: I18n.t('administration.assessor.assessments.bulk_delete_confirmation'),
+      okText: I18n.t('common.text.ok'),
+      cancelText: I18n.t('common.text.cancel'),
+      onOk: async () => {
+        await bulkDelete(parsedCampaignId, parsedAssessorId, selectedIds)
+        await fetch(parsedCampaignId, parsedAssessorId, tableConfig)
+        message.success(I18n.t('administration.assessor.assessments.bulk_delete_successful'))
+      },
+    })
   }
 
   return (
